enh: move crypto and web3 utils from Engima to respective dedicated modules
